zone_for_ip fallback for 127.0.0.0/8 address family
This ensures trackers which accept clients on the same host
as the tracker ends up in the correct network zone.

Defining a network zone explicitly for 127.0.0.0/8 in
server_settings would not work well because trackers
may be running in multiple zones themselves, so instead
use the external IP of the host the tracker is running on.